+++ packages/zowe-explorer/__tests__/__unit__/configuration/Profiles.unit.test.ts
@@ -2656,55 +2656,3 @@ describe("Profiles unit tests - function showProfilesInactiveMsg", () => {
         );
     });
 });
-
-describe("Profiles unit tests - function resolveTypePromise", () => {
-    it("should resolve each deferred promise of matching profile type", () => {
-        createGlobalMocks();
-        Object.defineProperty(Profiles.getInstance(), "allProfiles", {
-            value: [
-                { name: "sestest", type: "ssh" },
-                { name: "profile1", type: "zosmf" },
-                { name: "profile2", type: "ssh" },
-            ],
-            configurable: true,
-        });
-        const extenderTypeReadySpy = jest.spyOn(ProfilesUtils.extenderTypeReady, "get");
-        ProfilesUtils.resolveTypePromise("ssh");
-        expect(extenderTypeReadySpy).toHaveBeenCalledTimes(2);
-    });
-    it("should not resolve any deferred promise", () => {
-        createGlobalMocks();
-        Object.defineProperty(Profiles.getInstance(), "allProfiles", {
-            value: [
-                { name: "sestest", type: "zosmf" },
-                { name: "profile1", type: "zosmf" },
-                { name: "profile2", type: "zosmf" },
-            ],
-            configurable: true,
-        });
-        const extenderTypeReadySpy = jest.spyOn(ProfilesUtils.extenderTypeReady, "get");
-        ProfilesUtils.resolveTypePromise("ssh");
-        expect(extenderTypeReadySpy).toHaveBeenCalledTimes(0);
-    });
-    it("should resolve an existing promise without setting it", () => {
-        createGlobalMocks();
-        Object.defineProperty(Profiles.getInstance(), "allProfiles", {
-            value: [
-                { name: "sestest", type: "ssh" },
-                { name: "profile1", type: "zosmf" },
-                { name: "profile2", type: "ssh" },
-            ],
-            configurable: true,
-        });
-        jest.spyOn(ProfilesUtils.extenderTypeReady, "has").mockReturnValue(true);
-        const extenderTypeReadySetSpy = jest.spyOn(ProfilesUtils.extenderTypeReady, "set");
-        const extenderTypeReadyGetSpy = jest.spyOn(ProfilesUtils.extenderTypeReady, "get").mockReturnValue({
-            resolve: jest.fn(),
-            reject: jest.fn(),
-        });
-
-        ProfilesUtils.resolveTypePromise("ssh");
-        expect(extenderTypeReadySetSpy).toHaveBeenCalledTimes(0);
-        expect(extenderTypeReadyGetSpy).toHaveBeenCalledTimes(2);
-    });
-});
